refactor: move graph helpers to utils module
